# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Crop AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voice/guidance:
    post:
      operationId: getVoiceGuidance
      tags: [voice]
      summary: Get voice guidance text for a disease and language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceGuidanceBody"
      responses:
        "200":
          description: Voice guidance text generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceGuidanceResponse"
...
components:
  schemas:
...
    VoiceGuidanceBody:
      type: object
      properties:
        disease:
          type: string
        treatment:
          type: string
        language:
          type: string
          description: "en | id | th | vi | ph | jv | su | ms"
        cropType:
          type: string
      required:
        - disease
        - treatment
        - language
        - cropType
    VoiceGuidanceResponse:
      type: object
      properties:
        text:
          type: string
        language:
          type: string
        languageName:
          type: string
      required:
        - text
        - language
        - languageName
...