# lib/api-zod/src/generated/types/voiceGuidanceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Crop AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceGuidanceBody {
  disease: string;
  treatment: string;
  /** en | id | th | vi | ph | jv | su | ms */
  language: string;
  cropType: string;
}